Bug 1743834 - Fix static-analysis warning. r=alwu
Differential Revision: https://phabricator.services.mozilla.com/D141991

UltraBlame original commit: 322ac397ad9c0b0abc76a6471505202ab0e07ea0

+++ dom/media/mediasink/AudioSink.cpp
@@ -342,7 +342,8 @@ void AudioSink::NotifyAudioNeeded() {
              "Not called from the owner's thread");
 
   while (mAudioQueue.GetSize() &&
-         AudioQueuedInRingBufferMS() < mProcessedQueueThresholdMS) {
+         AudioQueuedInRingBufferMS() <
+             static_cast<uint32_t>(mProcessedQueueThresholdMS)) {
     
     if (mAudioQueue.PeekFront()->Frames() >
         SampleToFrame(mProcessedSPSCQueue->AvailableWrite())) {
